Conflicts:
	src/libtiled/libtiled.qbs
	src/libtiled/map.cpp
	src/libtiled/map.h
	src/plugins/python/pythonbind.cpp
	src/tiled/tilestamp.cpp
	src/tiled/tilestampsdock.cpp

+++ src/tiled/changetileprobability.h
@@ -35,11 +35,11 @@ class ChangeTileProbability : public QUndoCommand
 public:
     ChangeTileProbability(TilesetDocument *tilesetDocument,
                           const QList<Tile*> &tiles,
-                          float probability);
+                          qreal probability);
 
     ChangeTileProbability(TilesetDocument *tilesetDocument,
                           const QList<Tile*> &tiles,
-                          const QList<float> &probabilities,
+                          const QList<qreal> &probabilities,
                           QUndoCommand *parent = nullptr);
 
     void undo() override { swap(); }
@@ -50,7 +50,7 @@ private:
 
     TilesetDocument *mTilesetDocument;
     QList<Tile*> mTiles;
-    QList<float> mProbabilities;
+    QList<qreal> mProbabilities;
 };
 
 } // namespace Internal
